% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lillardhaz-package.R
\docType{package}
\name{lillardhaz-package}
\alias{lillardhaz}
\alias{lillardhaz-package}
\title{lillardhaz: Simultaneous-Equations Hazard/Probit Models}
\description{
Fits two-equation simultaneous models in the style of Lillard (1993): a
pair of processes -- each a binary probit outcome or a continuous-time
hazard duration -- linked through a single correlation parameter between
their underlying error terms, estimated jointly by maximum likelihood via
a Gaussian copula. See \code{\link{fit_lillardhaz}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nobifukuda/lillardhaz}
  \item Report bugs at \url{https://github.com/nobifukuda/lillardhaz/issues}
}

}
\author{
\strong{Maintainer}: Nobutaka Fukuda \email{nobutaka.fukuda@tohoku.ac.jp}

Authors:
\itemize{
  \item Nobutaka Fukuda \email{nobutaka.fukuda@tohoku.ac.jp}
}

}
